worker/orca-agent/requirements.txt
anthropic>=0.8.1
python-dotenv>=1.0.0
pandas>=2.0.0
tiktoken>=0.5.2
pytest>=8.0.2
typing-extensions>=4.12.2
GitPython>=3.1.44
pygithub>=2.5.0
Flask>=3.0.0
requests>=2.32.0
cryptography>=42.0.0
gunicorn>=22.0.0
solders>=0.26.0
base58>=2.1.0
tenacity>=9.0.0
sqlmodel>=0.0.22
openai>=0.28.0
colorama>=0.4.6
pymongo>=4.11.0
PyNaCl>=1.5.0
setuptools>=78.1.0
prometheus_swarm>=0.3.4
prometheus_test>=0.1.9
